fix: 修复编译错误、新增 Redis 配置、补全 README
- 修复 service 层 Date  LocalDateTime、int  long 类型问题
- 修复 UpdateWrapper 导入路径（MyBatis-Plus 3.5.5）
- 修复 Result 字段 msg  message 规范命名
- 修复 pom.xml 显式声明 mysql-connector-java 版本
- 新增 RedisConfig 提供 RedisTemplate<String, Object> Bean
- 完善 .gitignore 排除 target/、.trae/ 等
- 新增 README.md 项目说明文档

+++ cms-backend/src/main/java/com/cms/service/impl/CompetitionServiceImpl.java
@@ -41,7 +41,7 @@ public class CompetitionServiceImpl extends ServiceImpl<CompetitionMapper, Compe
         Competition old = this.getById(id);
         if (old == null) throw new BusinessException("竞赛不存在");
 
-        int regCount = registrationMapper.selectCount(new LambdaQueryWrapper<Registration>()
+        long regCount = registrationMapper.selectCount(new LambdaQueryWrapper<Registration>()
                 .eq(Registration::getCompetitionId, id))
             + teamRegistrationMapper.selectCount(new LambdaQueryWrapper<TeamRegistration>()
                 .eq(TeamRegistration::getCompetitionId, id));
@@ -76,7 +76,7 @@ public class CompetitionServiceImpl extends ServiceImpl<CompetitionMapper, Compe
     public void deleteCompetition(Long id) {
         Competition c = this.getById(id);
         if (c == null) throw new BusinessException("竞赛不存在");
-        int count = registrationMapper.selectCount(new LambdaQueryWrapper<Registration>()
+        long count = registrationMapper.selectCount(new LambdaQueryWrapper<Registration>()
                 .eq(Registration::getCompetitionId, id))
             + teamRegistrationMapper.selectCount(new LambdaQueryWrapper<TeamRegistration>()
                 .eq(TeamRegistration::getCompetitionId, id));
